refactor(ui): Move schemes to their own directory
