+++ apps/clearance_ui/src/components/license_conclusions/BulkConclusionForm.tsx
@@ -37,8 +37,8 @@ import {
     TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useToast } from "@/components/ui/use-toast";
-import { patternGlobSchema } from "@/components/common/pattern_schema";
-import { concludedLicenseExpressionSPDXSchema } from "@/components/common/spdx_schema";
+import { patternGlobSchema } from "@/schemes/pattern_schema";
+import { concludedLicenseExpressionSPDXSchema } from "@/schemes/spdx_schema";
 import ConclusionLicense from "@/components/license_conclusions/ConclusionLicense";
 import ConclusionSPDX from "@/components/license_conclusions/ConclusionSPDX";
 import { findMatchingPaths } from "@/helpers/findMatchingPaths";
