+++ apps/clearance_ui/src/components/license_conclusions/ConclusionForm.tsx
@@ -28,7 +28,7 @@ import {
     TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useToast } from "@/components/ui/use-toast";
-import { concludedLicenseExpressionSPDXSchema } from "@/components/common/spdx_schema";
+import { concludedLicenseExpressionSPDXSchema } from "@/schemes/spdx_schema";
 import ConclusionLicense from "@/components/license_conclusions/ConclusionLicense";
 import ConclusionSPDX from "@/components/license_conclusions/ConclusionSPDX";
 import { toPathPurl } from "@/helpers/pathParamHelpers";
