+++ apps/clearance_ui/src/components/main_ui/package_bulk_conclusions/BulkConclusionEditForm.tsx
@@ -33,11 +33,11 @@ import {
     TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { useToast } from "@/components/ui/use-toast";
-import { patternGlobSchema } from "@/components/common/pattern_schema";
-import { concludedLicenseExpressionSPDXSchema } from "@/components/common/spdx_schema";
 import ConclusionLicense from "@/components/license_conclusions/ConclusionLicense";
 import ConclusionSPDX from "@/components/license_conclusions/ConclusionSPDX";
 import { cn } from "@/lib/utils";
+import { patternGlobSchema } from "@/schemes/pattern_schema";
+import { concludedLicenseExpressionSPDXSchema } from "@/schemes/spdx_schema";
 
 const bulkConclusionFormSchema = z.object({
     pattern: patternGlobSchema,
