+++ apps/clearance_ui/src/schemes/pattern_schema.ts
