+++ apps/clearance_ui/src/schemes/spdx_schema.ts
